Merged removed changes. Added all code which should work except for the SINTIUM library. The previous commits contained multiple flaws which I have not fixed.
The integration needs to be organized as the code was, and since there were multiple redundant css files, I moved it to a css folder, so the refs needs tobe updated accordingly.

Furthermore, There should be only one js path from the root asset folder, so that if you want a segregated app, it should only contain the html and specialized js file, not multiple files which should reside in the js directory.

So whoever has an ide needs to integrate a MINIFIED and obfuscated of the lib. I would also like to see minified css&js files accross so that the map app simply needs to load one file.